Rename ensemble_model to ensemble_manager and tide_train to tide_model_trainer
Per PR #933 review: the service module, binary, cargo feature, devenv
commands, and log file all use ensemble_manager; the trainer binary is
tide_model_trainer. The fetch_equity_*_auto helpers become
fetch_equity_*_from_pool_or_service for clarity, and the duplicate
/model/predictions route is removed (nothing calls it; the portfolio
manager reads predictions from PostgreSQL, not HTTP).
