+++ src/bin/ensemble_manager.rs
@@ -1,0 +1,4 @@
+#[tokio::main]
+async fn main() {
+    fund::ensemble_manager::run("0.0.0.0:8082").await;
+}
